Move utilities to pkg
