+++ pkg/keystorev3/scrypt.go
@@ -21,8 +21,8 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/hyperledger/firefly-signer/internal/secp256k1"
 	"github.com/hyperledger/firefly-signer/internal/types"
+	"github.com/hyperledger/firefly-signer/pkg/secp256k1"
 	"github.com/hyperledger/firefly/pkg/fftypes"
 	"golang.org/x/crypto/scrypt"
 )
