+++ pkg/keystorev3/wallet_test.go
