+++ pkg/keystorev3/walletfile.go
@@ -20,8 +20,8 @@ import (
 	"bytes"
 	"fmt"
 
-	"github.com/hyperledger/firefly-signer/internal/secp256k1"
 	"github.com/hyperledger/firefly-signer/internal/types"
+	"github.com/hyperledger/firefly-signer/pkg/secp256k1"
 	"github.com/hyperledger/firefly/pkg/fftypes"
 )
